Updated data preparation for training
Added the core functionality of image reconstruction. The results seem consistent with the matlab code, but not explicit unit tests written.

In the current state this does not yet work for more angles than 1, need to see what the best way is to make it suitable for multiple angles. This might require modifying the deepus library functions which we should be very careful with.

Furthermore saving still needs to be implemented, and made compatible with loading in deepus functionality.